+++ NEWS.md
@@ -1,5 +1,7 @@
 # vroom (development version)
 
+* `vroom_write()` now works with all ALTREP vectors, including string vectors (#270)
+
 # vroom 1.3.2
 
 * Test failures on R 4.1 related to factors with NA values fixed (#262)